+++ packages/gamut-icons/src/types.ts
@@ -1,9 +1,12 @@
 import * as React from 'react';
 
-export interface GamutIconProps extends React.SVGProps<SVGSVGElement> {
+export type GamutIconProps = {
+  /**
+   * A suffix added to the end of the unique generated ID for the icon. This is useful if you have multiple of the same icon on the page and need to pass accessibility guidelines.
+   */
   titleId?: string;
   size?: number | string;
   title?: string;
   color?: string;
-  svgRef?: React.Ref<SVGSVGElement>;
-}
+  ref?: React.Ref<SVGSVGElement>;
+} & React.SVGProps<SVGSVGElement>;